版本时间:2015.08.21

+++ rd/trunk/Client/cms/Assets/script/Battle/BattleScript/NormalScript.cs
@@ -58,14 +58,14 @@ public static class NormalScript
                     maxLife += itor.Current.Value.maxHp;
                 }
 
-                lifeRatio = curLife / maxLife;
+                lifeRatio = (float)curLife / maxLife;
             }
             else 
             {
                 WeakPointRuntimeData wpRuntionData;
                 if (wpRutimeDataList.TryGetValue(wpName, out wpRuntionData))
                 {
-                    lifeRatio = wpRuntionData.hp / wpRuntionData.maxHp;
+                    lifeRatio = (float)wpRuntionData.hp / wpRuntionData.maxHp;
                 }
             }
         }
